Create Practical 9-Indivudal Avitivty-R Studio codes(Gene_expression and Gene_data)
this file contains codes which related to R studio. in this files consists of two parts with different codes part A consist of codes which used to download a .tsv file from the givel site and read the table properly. further more in this part there are codes which related to add a new row and regarding other mathematical operations such as find the row mean, ordering the values according to the descending order ,counting the number of genes which follow the given instruction and creating pairs plot.

part B consists of codes which are related to downloadeing .csv file from the given site and how to calculate the mean and the SD raw separately using subset codes. further more this files contains of codes of how to create a boxplot using the given columns.